+++ examples/car/main.py
@@ -297,6 +297,8 @@ def main() -> None:
     parser.add_argument("--save", type=str, default=None, help="Save data to a .rrd file at this path")
     args = parser.parse_args()
 
+    rerun.init("car")
+
     if args.serve:
         rerun.serve()
     elif args.connect: